refactor: 객체 역할을 내포하는 네이밍으로 수정
