+++ src/main/java/kr/allcll/backend/admin/seat/SeatStreamStatus.java
@@ -1,6 +1,6 @@
-package kr.allcll.backend.support.sse;
+package kr.allcll.backend.admin.seat;
 
-public enum SseStatus {
+public enum SeatStreamStatus {
     LIVE("과목의 여석을 실시간으로 볼 수 있어요."),
     PRESEAT("전체 학년 여석 탭을 이용해 주세요."),
     IDLE("과목 실시간 여석 제공 전이에요. 서비스가 시작되면 알림을 드릴게요."),
@@ -8,7 +8,7 @@ public enum SseStatus {
 
     private final String message;
 
-    SseStatus(String message) {
+    SeatStreamStatus(String message) {
         this.message = message;
     }
